+++ Packages/Accelib/Module/Accelib.Conditional/Runtime/Data/EValueSourceType.cs
@@ -8,6 +8,7 @@ namespace Accelib.Conditional.Data
         Integer = 0,
         Double = 1,
         Boolean = 2,
+        String = 3,
         [InspectorName("Preset")] ScriptableObject = 10,
         [InspectorName("Preset(Advanced)")] Custom = 11,
     }